fixed issues with adding and removing values from the default issueTypeScheme

+++ packages/jira-adapter/src/filters/issue_type_schemas/issue_type_scheme.ts
@@ -17,6 +17,7 @@ import { AdditionChange, CORE_ANNOTATIONS, getChangeData, InstanceElement, isAdd
 import { resolveChangeElement } from '@salto-io/adapter-utils'
 import _ from 'lodash'
 import { promises } from '@salto-io/lowerdash'
+import { logger } from '@salto-io/logging'
 import { getLookUpName } from '../../reference_mapping'
 import JiraClient from '../../client/client'
 import { JiraConfig } from '../../config'
@@ -26,6 +27,8 @@ import { getDiffIds } from '../../diff'
 
 const ISSUE_TYPE_SCHEMA_NAME = 'IssueTypeScheme'
 const MAX_CONCURRENT_PROMISES = 20
+
+const log = logger(module)
 
 const deployNewAndDeletedIssueTypeIds = async (
   change: ModificationChange<InstanceElement>,
@@ -38,21 +41,29 @@ const deployNewAndDeletedIssueTypeIds = async (
 
   const instance = getChangeData(change)
   if (addedIds.length > 0) {
-    await client.put({
-      url: `/rest/api/3/issuetypescheme/${instance.value.id}/issuetype`,
-      data: {
-        issueTypeIds: Array.from(addedIds),
-      },
-    })
+    if (!instance.value.isDefault) {
+      await client.put({
+        url: `/rest/api/3/issuetypescheme/${instance.value.id}/issuetype`,
+        data: {
+          issueTypeIds: Array.from(addedIds),
+        },
+      })
+    } else {
+      log.info('Skipping adding issues to default issue type scheme because they are automatically added')
+    }
   }
 
-  await promises.array.withLimitedConcurrency(
-    Array.from(removedIds).map(id => () =>
-      client.delete({
-        url: `/rest/api/3/issuetypescheme/${instance.value.id}/issuetype/${id}`,
-      })),
-    MAX_CONCURRENT_PROMISES,
-  )
+  if (!instance.value.isDefault) {
+    await promises.array.withLimitedConcurrency(
+      Array.from(removedIds).map(id => () =>
+        client.delete({
+          url: `/rest/api/3/issuetypescheme/${instance.value.id}/issuetype/${id}`,
+        })),
+      MAX_CONCURRENT_PROMISES,
+    )
+  } else if (removedIds.length > 0) {
+    log.info('Removing issue types from default issue type scheme is not supported')
+  }
 }
 
 const deployIssueTypeIdsOrder = async (
